refactor styles to make them more consistent
