+++ apps/search-poc/src/components/ResourceDetails/MorphologyViewer.css
@@ -1,4 +1,3 @@
-
-.morphology {
+.morphology-viewer {
   height: 400px;
 }
