+++ apps/search-poc/src/components/ResourceDetails/ResourceDetails.css
@@ -5,9 +5,6 @@
   justify-content: space-between;
 }
 
-.metadata-container {
-}
-
 .metadata-container > ul {
   padding: 0;
   margin: 0;
