+++ apps/search-poc/src/index.css
@@ -2,87 +2,16 @@ body {
   background-color: #f1f4f8;
 }
 
-.App {
+/* React App Root */
+#root {
   font-family: sans-serif;
   display: flex;
   flex-flow: row wrap;
-}
-
-.Header {
-  flex: 1 100%;
-  display: flex;
-  justify-content: space-between;
-  margin: 0.5em 1em;
-}
-
-.Header > .title {
-  font-weight: lighter;
-}
-.Header_logo {
-  width: 3em;
-  height: 3em;
-  position: absolute;
-  left: 50%;
-}
-.Header_auth {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  flex-grow: 1;
-  width: 3em;
-  height: 3em;
-}
-
-.ResultTable {
-  padding: 0 1em;
-  width: 100%;
-}
-
-.Filters {
-  min-width: 200px;
-  min-height: 200px;
-}
-
-.Results {
-  min-width: 200px;
-  flex-grow: 1;
-}
-
-.Dashboard-list > .item {
-  cursor: pointer;
-}
-
-.Dashboard-list > .item.active {
-  color: hotpink;
-}
-
-.ResourceDetails {
-  width: 980px;
-  margin: 0 auto;
-  display: flex;
-  justify-content: space-between;
-}
-
-.ResourceDetails > .metadata-container > ul {
-  padding: 0;
-  margin: 0;
-}
-.ResourceDetails > .metadata-container > .name {
-  font-size: 3em;
-  font-weight: lighter;
-}
-
-.small-table td {
-  padding: 4px 16px !important;
+  padding: 1em;
 }
 
 .ellipsis {
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
-}
-
-.main-view {
-  display: flex;
-  width: 100%;
 }
