+++ apps/search-poc/src/views/DetailsView.tsx
