+++ apps/search-poc/src/views/MainView.css
@@ -1,0 +1,4 @@
+.main-view {
+  display: flex;
+  width: 100%;
+}
